requirements.txt
django>=3.2
django_crispy_forms
Django-Verify-Email
django-decouple
social-auth-app-django